requirements.txt
whisperx>=0.10.1
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
